refactor(api): neutralize chat event and title service filenames
Rename services/zero-chat-events.command.ts to services/chat-events.command.ts
and services/zero-chat-title.service.ts to services/chat-title.service.ts.
Both files move with git mv and no declaration renaming; the only content edit
inside the pair is the intra-pair import line.

Closes #27989
