+++ turbo/apps/api/src/signals/services/shared-thread.service.ts
@@ -19,7 +19,7 @@ import {
 import { db$, writeDb$, type Db } from "../external/db";
 import { publishUserSignal } from "../external/realtime";
 import { visibleChatEventCondition } from "./chat-event-shared.service";
-import { generateSharedThreadTitle } from "./zero-chat-title.service";
+import { generateSharedThreadTitle } from "./chat-title.service";
 import { projectUserMessageForPublicShare } from "./chat-user-message.service";
 import {
   canonicalArchivedChatEventContent,
