+++ turbo/packages/db/scripts/agent-compose-consolidation-preflight-manifest.ts
@@ -177,6 +177,7 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/agentphone.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/agentphone.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/browser.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
+    "turbo/apps/api/src/signals/services/chat-events.command.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/chat-goal-queue.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/chat-search.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/chat-thread-event.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
@@ -227,7 +228,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/webhooks-clerk-cleanup.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/workflow-automation.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/workflow-data.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-chat-events.command.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/packages/db/src/schema/agent-draft.ts|./zero-agent|zeroAgents:zeroAgents",
@@ -290,12 +290,14 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/browser.service.ts|agentComposeId,agentComposes",
     "turbo/apps/api/src/signals/services/canonical-feishu-ingress-processor.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/chat-event-shared.service.ts|agentComposeId",
+    "turbo/apps/api/src/signals/services/chat-events.command.ts|agentComposeId,zeroAgents",
     "turbo/apps/api/src/signals/services/chat-goal-queue.service.ts|agentComposeId,zeroAgents",
     "turbo/apps/api/src/signals/services/chat-search.service.ts|agentComposeId,agentComposes",
     "turbo/apps/api/src/signals/services/chat-session-continuity.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/chat-thread-event.service.ts|agentComposeId,agentComposes",
     "turbo/apps/api/src/signals/services/chat-thread-model.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/chat-thread.service.ts|agentComposeId,zeroAgents",
+    "turbo/apps/api/src/signals/services/chat-title.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/compose-data.service.ts|agentComposeId,agentComposeVersions,agentComposes",
     "turbo/apps/api/src/signals/services/computer-use-authorization.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/computer-use.service.ts|agentComposeId",
@@ -355,8 +357,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/workflow-data.service.ts|zeroAgents",
     "turbo/apps/api/src/signals/services/workflow-user-automation-thread.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/workflow-volume.service.ts|headVersionId",
-    "turbo/apps/api/src/signals/services/zero-chat-events.command.ts|agentComposeId,zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-chat-title.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|agentComposeId,agentComposeVersions,agentComposes,headVersionId,zeroAgents",
     "turbo/apps/platform/src/views/zero-page/zero-telegram-settings-page.tsx|headVersionId",
     "turbo/packages/api-contracts/src/contracts/composes.ts|headVersionId",
